created vagrant guest plugin for rancheros
