+++ vagrant-rancheros/Gemfile
@@ -1,0 +1,9 @@
+source 'https://rubygems.org'
+
+group :development do
+  gem "vagrant", :git => "git://github.com/mitchellh/vagrant.git"
+end
+
+group :plugins do
+  gem "vagrant-rancheros", :path => "."
+end
