+++ vagrant-rancheros/lib/vagrant-rancheros.rb
@@ -1,0 +1,1 @@
+require "vagrant-rancheros/plugin"
