+++ vagrant-rancheros/lib/vagrant-rancheros/version.rb
@@ -1,0 +1,5 @@
+module VagrantPlugins
+  module GuestRancheros
+    VERSION = "0.0.1"
+  end
+end
